+++ src/components/Modals/DeleteRelationship/DeleteRelationshipModal.jsx
@@ -1,56 +1,19 @@
-import { useSetAtom, useAtomValue } from "jotai";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useSetAtom } from "jotai";
 
-import fetchData from "../../../utils/axios";
-import {
-  currentDBIdAtom,
-  userAtom,
-  showDeleteRelationshipModalAtom,
-  deleteTargetRelationshipAtom,
-} from "../../../atoms/atoms";
+import { showDeleteRelationshipModalAtom } from "../../../atoms/atoms";
+import useMutateDeleteRelationship from "../../../apis/useMutateDeleteRelationship";
 
 import Modal from "../../shared/Modal";
 import Button from "../../shared/Button";
 import Content from "../SharedItems/Content";
 import ContentWrapper from "../SharedItems/ContentWrapper";
 import Message from "../SharedItems/Message";
-import Loading from "../../shared/Loading";
 
 function DeleteRelationshipModal() {
-  const queryClient = useQueryClient();
-
-  const { userId } = useAtomValue(userAtom);
-  const currentDBId = useAtomValue(currentDBIdAtom);
-  const deleteTargetRelationship = useAtomValue(deleteTargetRelationshipAtom);
-
   const setShowDeleteRelationshipModal = useSetAtom(
     showDeleteRelationshipModalAtom,
   );
-
-  async function deleteRelationship() {
-    await fetchData(
-      "DELETE",
-      `/users/${userId}/databases/${currentDBId}/relationships/${deleteTargetRelationship}`,
-    );
-  }
-
-  const { mutate: fetchDeleteRelationship, isLoading } = useMutation(
-    deleteRelationship,
-    {
-      onSuccess: () => {
-        queryClient.refetchQueries(["dbDocumentList", currentDBId]);
-        queryClient.refetchQueries(["dbRelationShips", currentDBId]);
-        setShowDeleteRelationshipModal(false);
-      },
-      onFailure: () => {
-        console.log("sending user to errorpage");
-      },
-    },
-  );
-
-  if (isLoading) {
-    return <Loading />;
-  }
+  const fetchDeleteRelationship = useMutateDeleteRelationship();
 
   return (
     <Modal onClick={() => setShowDeleteRelationshipModal(false)}>
